services/ai-llm/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
anthropic>=0.52.0
google-genai>=1.0.0
openai>=1.50.0
pydantic>=2.0.0
python-dotenv>=1.0.0
redis>=5.0.0
httpx>=0.27.0
